+ provide ip-based access for git users
~/.ssh/authorized_keys supports :from field to specify
locations from that user can access to :ssh server. This
feature is to provide this ability to :gitolite, and it
is written in Ruby (sorry, I am not good at Perl). The
script reads configuration from
  ~/.ssh/gitolite.users.address
(which is a YAML file) and update ~/.ssh/authorized_keys.
The configuration is as in this example
  user-foo:
  - x.y.y.z
  - example.com
  user-bar:
  - x.x.y.z
  - example.bar
The configuration file path is hard-coded